quirks: Adds a u-boot quirk
What does this mean?

This means that it's possible to get a *full* disk image for a u-boot
using target. That full disk image can be flashed to a storage, e.g. an
SD card, and should be fully bootable.

It is expected that such u-boot using targets would be using u-boots
with android-like configurations. Mainly, reading the boot.img and
booting it.

At the moment, this may be requiring a couple patches to u-boot.